pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "project-ult-subsystem-holdings"
version = "0.1.0"
description = "Holdings producer subsystem MVP scaffold for project-ult."
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "pydantic>=2.6",
    "project-ult-contracts>=0.1.3",
    "project-ult-subsystem-sdk>=0.1.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]
contracts-schemas = [
    "project-ult-contracts @ git+https://github.com/shenfanjie5-bit/project-ult-contracts.git@v0.1.3",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["subsystem_holdings*"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
pythonpath = ["src"]
